+++ pkg/privatetypes/t_url.go
@@ -16,7 +16,7 @@ func init() {
 	Register(TypeURL, "URL", func() dnsv2.RR { return new(URL) }, privatetypesrdata.MakeURL)
 }
 
-const TypeURL = 65299
+const TypeURL = uint16(65299)
 
 type URL struct {
 	Hdr dnsv2.Header
